moved wire_node to common as well. Hopeffuly last dependency
