+++ include/coreir/common/op_graph.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "coreir/simulator/wire_node.h"
+#include "coreir/common/wire_node.h"
 
 namespace CoreIR {
