+++ include/coreir/common/wire_node.h
